improved tox testing, improved docstrings all over. All doc tests pass.

+++ kecpkg/utils.py
@@ -8,21 +8,43 @@ from kecpkg.commands.utils import echo_warning, echo_failure, echo_info
 
 
 def ensure_dir_exists(d):
+    """Ensure that directory exists, otherwise make directory."""
     if not os.path.exists(d):
         os.makedirs(d)
 
 
-def create_file(filepath, content=None):
+def create_file(filepath, content=None, overwrite=True):
+    """
+    Create file and optionally fill it with content.
+
+    Will overwrite file already in place if overwrite flag is set
+
+    :param filepath: full path to a file to create
+    :param content:
+    :return:
+    """
     ensure_dir_exists(os.path.dirname(os.path.abspath(filepath)))
-    with open(filepath, 'w') as fd:
-        os.utime(filepath, times=None)
-        if content:
-            fd.write(content)
+    # if overwrite is set to True overwrite file, otherwise if file exist, exit.
+
+    if not os.path.exists(filepath) or (os.path.exists(filepath) and overwrite):
+        with open(filepath, 'w') as fd:
+            os.utime(filepath, times=None)
+            if content:
+                fd.write(content)
+    else:
+        echo_failure("File '{}' already exists.".format(filepath))
+        sys.exit(1)
 
 
-def download_file(url, fname):
+def download_file(url, filepath):
+    """
+    Download file from url and save to disk as filename.
+
+    :param url: url to download file from
+    :param filepath: filename to write to
+    """
     req = urlopen(url)
-    with open(fname, 'wb') as f:
+    with open(filepath, 'wb') as f:
         while True:
             chunk = req.read(16384)
             if not chunk:
@@ -31,18 +53,29 @@ def download_file(url, fname):
             f.flush()
 
 
-def copy_path(path, d):
-    if os.path.isdir(path):
+def copy_path(sourcepath, destpath):
+    """
+    Copy path.
+
+    :param sourcepath: source path to copy from, if dir, copy subtree
+    :param destpath: destination path to copy to
+    """
+    if os.path.isdir(sourcepath):
         shutil.copytree(
-            path,
-            os.path.join(d, basepath(path)),
+            sourcepath,
+            os.path.join(destpath, basepath(sourcepath)),
             copy_function=shutil.copy
         )
     else:
-        shutil.copy(path, d)
+        shutil.copy(sourcepath, destpath)
 
 
 def remove_path(path):
+    """
+    Remove directory structure.
+
+    :param path: path to remove
+    """
     try:
         shutil.rmtree(path)
     except (FileNotFoundError, OSError):
@@ -53,14 +86,29 @@ def remove_path(path):
 
 
 def basepath(path):
+    """Get full basepath from path."""
     return os.path.basename(os.path.normpath(path))
 
 
-def normalise_name(package_name):
-    return re.sub(r"[-_. ]+", "_", package_name).lower()
+def normalise_name(raw_name):
+    """
+    Normalise the name to be used in python package allowable names.
+
+    conforms to PEP-423 package naming conventions
+
+    :param raw_name: raw string
+    :return: normalised string
+    """
+    return re.sub(r"[-_. ]+", "_", raw_name).lower()
 
 
 def get_package_dir(package_name=None):
+    """
+    Check and retrieve the package directory.
+
+    :param package_name: (optional) package name
+    :return: full path name to the package directory
+    """
     package_dir = package_name and os.path.join(os.getcwd(), package_name) or os.getcwd()
 
     try:
@@ -81,7 +129,7 @@ def get_package_dir(package_name=None):
 
 def get_artifacts_on_disk(root_path, exclude_paths=('venv', 'dist'), verbose=False):
     """
-    retrieve all artifacts on disk
+    Retrieve all artifacts on disk.
 
     :param root_path: root_path to collect all artifacts from
     :param exclude_paths: (optional) directory names and filenames to exclude
